Moved the ClaimRequirements to the helpers folder.
